[components testing] Refactor testing module
