+++ python_modules/dagster/dagster/components/testing/__init__.py
@@ -1,0 +1,16 @@
+from dagster.components.testing.deprecated.utils import (
+    scaffold_defs_sandbox as scaffold_defs_sandbox,
+)
+from dagster.components.testing.test_cases import (
+    TestOpCustomization as TestOpCustomization,
+    TestTranslation as TestTranslation,
+    TestTranslationBatched as TestTranslationBatched,
+)
+from dagster.components.testing.utils import (
+    copy_code_to_file as copy_code_to_file,
+    get_all_components_defs_from_defs_path as get_all_components_defs_from_defs_path,
+    get_module_path as get_module_path,
+    get_original_module_name as get_original_module_name,
+    random_importable_name as random_importable_name,
+    temp_components_sandbox as temp_components_sandbox,
+)
